requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Machine Learning
scikit-learn==1.3.2
imbalanced-learn==0.11.0
joblib==1.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
mysql-connector-python==8.2.0
python-dotenv==1.0.0

# LLM Integration
requests==2.31.0
PyYAML==6.0.1